fix: number rounding

+++ kata/6 kyu/children-and-candies/main/CircleOfChildren.java
@@ -5,7 +5,7 @@ class CircleOfChildren {
     int times = 0, dist[];
     while (of(dist = candies.clone()).distinct().count() != 1) {
       for (int i = 0; i < dist.length; i++) {
-        candies[i] += Math.ceil((i > 0 ? dist[i - 1] : dist[dist.length - 1]) / 2.) - dist[i] / 2.;
+        candies[i] += (int) Math.ceil((i > 0 ? dist[i - 1] : dist[dist.length - 1]) / 2.) - dist[i] / 2;
       }
       times++;
     }
